Factor out ImportAliasesDisplayColumnFactory to its own class
