+++ experiment/src/org/labkey/experiment/api/ExpSampleTypeTableImpl.java
@@ -16,23 +16,11 @@
 
 package org.labkey.experiment.api;
 
-import com.fasterxml.jackson.core.util.DefaultIndenter;
-import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.jetbrains.annotations.NotNull;
-import org.json.JSONObject;
 import org.labkey.api.data.AbstractTableInfo;
-import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.ContainerFilter;
-import org.labkey.api.data.DataColumn;
-import org.labkey.api.data.DisplayColumn;
-import org.labkey.api.data.DisplayColumnFactory;
 import org.labkey.api.data.JdbcType;
 import org.labkey.api.data.MutableColumnInfo;
-import org.labkey.api.data.RenderContext;
 import org.labkey.api.data.SQLFragment;
-import org.labkey.api.exp.api.ExpSampleType;
-import org.labkey.api.exp.api.SampleTypeService;
 import org.labkey.api.exp.query.ExpSampleTypeTable;
 import org.labkey.api.query.AliasedColumn;
 import org.labkey.api.query.DetailsURL;
@@ -40,16 +28,10 @@ import org.labkey.api.query.ExprColumn;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.permissions.InsertPermission;
 import org.labkey.api.security.permissions.UpdatePermission;
-import org.labkey.api.util.HtmlString;
-import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.experiment.controllers.exp.ExperimentController;
 
-import java.io.IOException;
 import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import static org.labkey.api.exp.api.SampleTypeService.MATERIAL_INPUTS_PREFIX;
 
@@ -146,94 +128,5 @@ public class ExpSampleTypeTableImpl extends ExpTableImpl<ExpSampleTypeTable.Colu
         detailsURL.setContainerContext(_userSchema.getContainer());
         setDetailsURL(detailsURL);
         setImportURL(AbstractTableInfo.LINK_DISABLER);
-    }
-
-    static class ImportAliasesDisplayColumnFactory implements DisplayColumnFactory
-    {
-        private final String _prefix;
-
-        public ImportAliasesDisplayColumnFactory(String prefix)
-        {
-            _prefix = prefix;
-        }
-
-        @Override
-        public DisplayColumn createRenderer(ColumnInfo colInfo)
-        {
-            DataColumn dataColumn = new DataColumn(colInfo)
-            {
-                private JSONObject getValueFromCtx(RenderContext ctx) throws IOException
-                {
-                    JSONObject json = null;
-                    Integer rowId = (Integer)getValue(ctx);
-                    ExpSampleType sampleType = SampleTypeService.get().getSampleType(rowId);
-                    if (sampleType != null)
-                    {
-                        Map<String, String> aliasMap = sampleType.getImportAliasMap();
-                        List<String> importKeys = aliasMap.keySet().stream()
-                                .filter(key -> aliasMap.get(key).startsWith(_prefix))
-                                .sorted()
-                                .collect(Collectors.toList());
-
-                        if (importKeys.size() > 0)
-                        {
-                            json = new JSONObject();
-                            for (String importKey : importKeys)
-                                json.put(importKey, aliasMap.get(importKey).substring(_prefix.length()));
-                        }
-                    }
-
-                    return json;
-                }
-
-                @Override
-                public Object getJsonValue(RenderContext ctx)
-                {
-                    Object value = getDisplayValue(ctx);
-                    return value != null ? value.toString() : null;
-                }
-
-                @Override
-                public Object getDisplayValue(RenderContext ctx)
-                {
-                    try
-                    {
-                        return getValueFromCtx(ctx);
-                    }
-                    catch (IOException e)
-                    {
-                        return HtmlString.of("Bad import alias object");
-                    }
-                }
-
-                @NotNull
-                @Override
-                public HtmlString getFormattedHtml(RenderContext ctx)
-                {
-                    try
-                    {
-                        JSONObject value = getValueFromCtx(ctx);
-                        if (null == value)
-                            return HtmlString.EMPTY_STRING;
-
-                        ObjectMapper mapper = new ObjectMapper();
-                        DefaultPrettyPrinter pp = new DefaultPrettyPrinter();
-                        pp.indentArraysWith(new DefaultIndenter());
-
-                        Object json = mapper.readValue(value.toString(), Object.class);
-                        String strValue = mapper.writer(pp).writeValueAsString(json);
-                        String filteredValue = PageFlowUtil.filter(strValue, true);
-                        return HtmlString.unsafe("<div>" + filteredValue + "</div>");
-                    }
-                    catch (IOException e)
-                    {
-                        return HtmlString.of("Bad import alias object");
-                    }
-                }
-            };
-
-            dataColumn.setTextAlign("left");
-            return dataColumn;
-        }
     }
 }
